+++ scripts/py2md.py
@@ -371,13 +371,17 @@ class AllenNlpRenderer(MarkdownRenderer):
             return signature
 
     def _format_classdef_signature(self, cls: Class) -> str:
+        code = ""
+        if cls.decorators:
+            for dec in cls.decorators:
+                code += "@{}{}\n".format(dec.name, dec.args or "")
         bases = ", ".join(map(str, cls.bases))
         if cls.metaclass:
             bases += ", metaclass=" + str(cls.metaclass)
         if bases:
-            code = "class {}({})".format(cls.name, bases)
+            code += "class {}({})".format(cls.name, bases)
         else:
-            code = "class {}".format(cls.name)
+            code += "class {}".format(cls.name)
         if self.signature_python_help_style:
             code = cls.path() + " = " + code
         if self.classdef_render_init_signature_if_needed and (